feat: move collectorProfile field to own file
