+++ src/schema/v2/me/index.ts
@@ -28,7 +28,7 @@ import Bidders from "./bidders"
 import { BidderPosition } from "./bidder_position"
 import BidderPositions from "./bidder_positions"
 import BidderStatus from "./bidder_status"
-import CollectorProfile from "./collector_profile"
+import { CollectorProfile } from "../CollectorProfile/collectorProfile"
 import Conversation from "./conversation"
 import Invoice from "./conversation/invoice"
 import Conversations from "./conversations"
